requirements.txt
numpy
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
fastapi
uvicorn[standard]
python-multipart
